requirements.txt
python-telegram-bot==20.6
google-generativeai==0.5.4
python-dotenv==1.0.1